src/main/java/com/alianz/practice/alianz_practice/config/JWTAuthenticationFilter.java
package com.alianz.practice.alianz_practice.config;

import java.io.IOException;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
import org.springframework.stereotype.Component;
import org.springframework.web.filter.OncePerRequestFilter;

import jakarta.annotation.Nonnull;
import jakarta.servlet.FilterChain;
import jakarta.servlet.ServletException;
import jakarta.servlet.http.Cookie;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;
import lombok.RequiredArgsConstructor;

@Component
@RequiredArgsConstructor
public class JWTAuthenticationFilter extends OncePerRequestFilter{
    @Autowired
    private JWTService jwtService;

    @Autowired
    private UserDetailsService userDetailsService;

    private static final String AUTH_WHITELIST = "/alianz/auth/**";

    
    @Override
    protected void doFilterInternal(@Nonnull HttpServletRequest request,@Nonnull HttpServletResponse response,@Nonnull FilterChain filterChain)
            throws ServletException, IOException {
       Cookie[] cookies = request.getCookies();
         String token = null;
         if(new AntPathRequestMatcher(AUTH_WHITELIST).matches(request)){
             filterChain.doFilter(request, response);
             return;
         }
         if(cookies == null){
                filterChain.doFilter(request, response);
                return;
         }
         for(Cookie cookie : cookies){
             if(cookie.getName().equals("token")){
                 token = cookie.getValue();
                    break;
             }
         }
         if(token == null){
             filterChain.doFilter(request, response);
             return;
         }
         final String username = jwtService.extractUser(token);
         if(username != null && SecurityContextHolder.getContext().getAuthentication() == null){
            UserDetails userDetails = this.userDetailsService.loadUserByUsername(username);
            if(jwtService.isTokenValid(token, userDetails)){
                UsernamePasswordAuthenticationToken authToken = new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
                authToken.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
                SecurityContextHolder.getContext().setAuthentication(authToken);
            }
         }
         filterChain.doFilter(request, response);
    }
    
}
